# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/tenant-availability": {
      "get": {
        "operationId": "AuthenticationController_tenantAvailability",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "query",
            "schema": {
              "example": "acme-pharma",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantAvailabilityResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/register-company": {
      "post": {
        "operationId": "AuthenticationController_registerCompany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthenticationController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Credentials are invalid."
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthenticationController_refresh",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "qualyra_refresh": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthenticationController_logout",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "qualyra_refresh": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "operationId": "AuthenticationController_logoutAll",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "qualyra_refresh": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthenticationController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "operationId": "AuthenticationController_listSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/sessions/{sessionId}": {
      "delete": {
        "operationId": "AuthenticationController_revokeSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthenticationController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeutralResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthenticationController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeutralResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "operationId": "AuthenticationController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeutralResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/resend-verification": {
      "post": {
        "operationId": "AuthenticationController_resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeutralResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Qualyra GMP API",
    "description": "Secure multi-tenant quality management API.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "qualyra_refresh"
      }
    },
    "schemas": {
...
      "TenantAvailabilityResponseDto": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean"
          }
        },
        "required": [
          "available"
        ]
      },
      "RegisterCompanyDto": {
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "example": "ACME Pharma"
          },
          "tenantSlug": {
            "type": "string",
            "example": "acme-pharma"
          },
          "adminName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "example": "jane@acme.example"
          },
          "password": {
            "type": "string",
            "example": "a long passphrase",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "tenantName",
          "tenantSlug",
          "adminName",
          "email",
          "password"
        ]
      },
      "AuthenticatedUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "displayName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "INVITED",
              "ACTIVE",
              "LOCKED",
              "DISABLED"
            ]
          },
          "emailVerifiedAt": {
            "type": "object",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "status"
        ]
      },
      "TenantSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "AuthenticationResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Short-lived RS256 access token."
          },
          "csrfToken": {
            "type": "string",
            "description": "Double-submit token for cookie-based operations."
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUserDto"
          },
          "tenant": {
            "$ref": "#/components/schemas/TenantSummaryDto"
          }
        },
        "required": [
          "accessToken",
          "csrfToken",
          "user",
          "tenant"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "tenant": {
            "type": "string",
            "example": "acme-pharma"
          },
          "email": {
            "type": "string",
            "example": "jane@acme.example"
          },
          "password": {
            "type": "string",
            "example": "a long passphrase"
          }
        },
        "required": [
          "tenant",
          "email",
          "password"
        ]
      },
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUserDto"
          },
          "tenant": {
            "$ref": "#/components/schemas/TenantSummaryDto"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "user",
          "tenant",
          "roles",
          "permissions"
        ]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "device": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "isCurrent": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "REVOKED",
              "EXPIRED"
            ]
          }
        },
        "required": [
          "id",
          "device",
          "createdAt",
          "lastUsedAt",
          "expiresAt",
          "isCurrent",
          "status"
        ]
      },
      "TenantEmailDto": {
        "type": "object",
        "properties": {
          "tenant": {
            "type": "string",
            "example": "acme-pharma"
          },
          "email": {
            "type": "string",
            "example": "jane@acme.example"
          }
        },
        "required": [
          "tenant",
          "email"
        ]
      },
      "NeutralResponseDto": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "accepted"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Opaque one-time token received by email."
          },
          "newPassword": {
            "type": "string",
            "example": "a new long passphrase",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "TokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Opaque one-time token received by email."
          }
        },
        "required": [
          "token"
        ]
      },
...
    }
  }
}